refactor: Make multicolorTasks use uuid instead of player object, handle multicolortask remove from team and cancel task method.
