+++ src/main/java/hhitt/fancyglow/managers/GlowManager.java
@@ -1,20 +1,24 @@
-package hhitt.fancyglow.utils;
+package hhitt.fancyglow.managers;
 
 import hhitt.fancyglow.FancyGlow;
 import hhitt.fancyglow.tasks.MulticolorTask;
+import hhitt.fancyglow.utils.MessageUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.scoreboard.Scoreboard;
 import org.bukkit.scoreboard.Team;
 
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
+import java.util.UUID;
 
 public class GlowManager {
 
     private final FancyGlow plugin;
     private final Map<ChatColor, Team> glowTeams;
-    private final Map<Player, MulticolorTask> multicolorTasks;
+    private final Map<UUID, MulticolorTask> multicolorTasks;
 
     public GlowManager(FancyGlow plugin) {
         this.plugin = plugin;
@@ -23,7 +27,7 @@ public class GlowManager {
     }
 
     public void toggleMulticolorGlow(Player player) {
-        if (multicolorTasks.containsKey(player)) {
+        if (isMulticolorTaskActive(player)) {
             removeGlow(player);
             MessageUtils.miniMessageSender(player, plugin.getMainConfigManager().getDisableGlow());
             return;
@@ -32,48 +36,51 @@ public class GlowManager {
         int ticks = plugin.getConfig().getInt("Rainbow_Update_Interval");
         MulticolorTask task = new MulticolorTask(plugin, player);
         task.runTaskTimer(plugin, 0L, ticks);
-        multicolorTasks.put(player, task);
+        multicolorTasks.put(player.getUniqueId(), task);
         player.setGlowing(true);
         MessageUtils.miniMessageSender(player, plugin.getMainConfigManager().getEnableGlow());
     }
 
     public boolean isMulticolorTaskActive(Player player) {
-        return multicolorTasks.containsKey(player);
+        return multicolorTasks.containsKey(player.getUniqueId());
     }
 
     public void toggleGlow(Player player, ChatColor color) {
         Team glowTeam = getOrCreateTeam(color);
         String cleanName = ChatColor.stripColor(player.getName());
 
+        // Remove any existing glow
         removeGlow(player);
 
         if (glowTeam.hasEntry(cleanName)) {
+            // Send message if no longer glowing.
             MessageUtils.miniMessageSender(player, plugin.getMainConfigManager().getDisableGlow());
             return;
         }
 
+        // Add the player to the team and enable glowing
         glowTeam.addEntry(cleanName);
         player.setGlowing(true);
         MessageUtils.miniMessageSender(player, plugin.getMainConfigManager().getEnableGlow());
     }
 
     public void removeGlow(Player player) {
+        // Remove glow from player.
         player.setGlowing(false);
+        // Remove from any existing color team.
         removePlayerFromAllTeams(player);
-
-        if (isMulticolorTaskActive(player)) {
-            multicolorTasks.get(player).cancel();
-            multicolorTasks.remove(player);
-        }
     }
 
     public void removePlayerFromAllTeams(Player player) {
         Scoreboard board = plugin.getServer().getScoreboardManager().getMainScoreboard();
         String cleanName = ChatColor.stripColor(player.getName());
-        for (Team team : board.getTeams()) {
-            if (team.hasEntry(cleanName)) {
-                team.removeEntry(cleanName);
-            }
+
+        // Handle remove if rainbow selected.
+        if (isMulticolorTaskActive(player)) {
+            // Cancel task
+            multicolorTasks.get(player.getUniqueId()).cancel();
+            // Remove from map.
+            multicolorTasks.remove(player.getUniqueId());
         }
 
         // Attempt to remove player from any color team.
@@ -105,5 +112,12 @@ public class GlowManager {
 
     public boolean hasGlowPermission(Player player, ChatColor color) {
         return player.hasPermission("fancyglow." + color.name().toLowerCase());
+    }
+
+    public void cancelMulticolorTasks() {
+        for (MulticolorTask task : multicolorTasks.values()) {
+            task.cancel();
+        }
+        multicolorTasks.clear();
     }
 }
